// src/app/models/grocery-categories.enum.ts
export enum GroceryCategoryEnum {
  FRUITS = 'Fruits',
  VEGETABLES = 'Vegetables',
  CANNED_GOODS = 'Canned Goods',
  DAIRY_AND_EGGS = 'Dairy & Eggs',
  BREAD_AND_BAKERY = 'Bread & Bakery',
  MEAT_AND_SEAFOOD = 'Meat & Seafood',
  FROZEN_FOODS = 'Frozen Foods',
  PANTRY_STAPLES = 'Pantry Staples',
  SNACKS = 'Snacks',
  BEVERAGES = 'Beverages',
  CONDIMENTS_AND_SAUCES = 'Condiments & Sauces',
  HOUSEHOLD_AND_CLEANING = 'Household & Cleaning',
  PERSONAL_CARE = 'Personal Care',
  BABY_ITEMS = 'Baby Items',
  PET_CARE = 'Pet Care'
}